Moved dotnet command to dedicated package
