+++ pkg/fleet/installer/packages/apm_library_dotnet_windows.go
@@ -13,7 +13,7 @@ import (
 	"io/fs"
 	"path/filepath"
 
-	"github.com/DataDog/datadog-agent/pkg/fleet/installer/exec"
+	"github.com/DataDog/datadog-agent/pkg/fleet/installer/packages/exec"
 	"github.com/DataDog/datadog-agent/pkg/fleet/installer/paths"
 	"github.com/DataDog/datadog-agent/pkg/fleet/installer/telemetry"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
